+++ NEWS.md
@@ -1,5 +1,7 @@
 # seaice dev
 
+* For NSIDC sources we now have argument `.local_root`, to override read from FTP and use the local files instead. 
+
 * Add AMSR2 3km files. 
 
 * Renamed functions to be more consistent, so `north/south_nsidc_*` functions are removed, in favour of `nsidc_south/north`.